Quick Start
This is a reference guide if you don't want to go through 16 hours of stream, this is a start of a readme, SQL at the bottom to bring tables triggers, function to current state happy-hour-18. This is a quick reference not perfect, but a good start. This should get through up to happy-hour-14 tables up to Happy days-#18. At the bottom is the complete SQL... Maybe some errors, I will update next week and test.